+++ src/typed_settings/converters.py
@@ -26,7 +26,7 @@ from typing import (
     get_origin,
 )
 
-from ._compat import PY_310
+from ._compat import PY_310, PY_311
 
 
 if PY_310:
@@ -49,7 +49,7 @@ if TYPE_CHECKING:
 #     "get_default_structure_hooks",
 #     "register_attrs_hook_factory",
 #     "register_strlist_hook",
-#     "to_dt",
+#     "to_datetime",
 #     "to_bool",
 #     "to_enum",
 #     "to_path",
@@ -187,7 +187,9 @@ def default_converter(*, resolve_paths: bool = True) -> Converter:
         - :class:`int`
         - :class:`float`
         - :class:`str`
-        - :class:`datetime.datetime` (see :func:`to_dt()`)
+        - :class:`datetime.datetime` (see :func:`to_datetime()`)
+        - :class:`datetime.date` (see :func:`to_date()`)
+        - :class:`datetime.timedelta` (see :func:`to_timedelta()`)
         - :class:`enum.Enum` using (see :func:`to_enum()`)
         - :class:`pathlib.Path` (see :func:`to_path()` and :func:`to_resolved_path()`)
         - :class:`list`
@@ -211,6 +213,8 @@ def default_converter(*, resolve_paths: bool = True) -> Converter:
     .. versionchanged:: 23.1.0
        Return a :program:`cattrs` converter if it is installed or else a Typed Settings
        converter.
+    .. versionchanged:: 24.3.0
+       Added :func:`to_date()` and :func:`to_timedelta()`.
     """
     try:
         import cattrs  # noqa: F401
@@ -516,35 +520,63 @@ def to_bool(value: Any, _cls: type = bool) -> bool:
     raise ValueError(f"Cannot convert value to bool: {value}")
 
 
-def to_date(value: Union[datetime, str], cls: type = date) -> date:
-    pass
-
-
-def to_datetime(value: Union[datetime, str], cls: type = datetime) -> datetime:
+def to_date(value: Union[datetime, str], cls: Type[date] = date) -> date:
     """
-    Convert an ISO formatted string to :class:`datetime.datetime`.  Leave the
-    input untouched if it is already a datetime.
+    Convert an ISO formatted string to :class:`datetime.date`.  Leave the input
+    untouched if it is already a date.
 
-    See: :meth:`datetime.datetime.fromisoformat()`
-
-    The ``Z`` suffix is also supported and will be replaced with ``+00:00``.
+    See: :meth:`datetime.date.fromisoformat()`
 
     Args:
         value: The input data
-        _cls: (ignored)
+        cls: The target type.  Must be :class:`datetime.date` or a subclass.
+
+    Return:
+        The converted date instance
+
+    Raise:
+        TypeError: If *value* is neither a string nor a date
+        ValueError: If *value* cannot be parsed as date.
+
+    .. versionadded:: 24.3.0
+    """
+    if not isinstance(value, (date, str)):
+        raise TypeError(
+            f"Invalid type {type(value).__name__!r}; expected 'date' or " f"'str'."
+        )
+    if isinstance(value, str):
+        return cls.fromisoformat(value)
+    return value
+
+
+def to_datetime(
+    value: Union[datetime, str], cls: Type[datetime] = datetime
+) -> datetime:
+    """
+    Convert an ISO formatted string to :class:`datetime.datetime`.  Leave the input
+    untouched if it is already a datetime.
+
+    See: :meth:`datetime.datetime.fromisoformat()`
+
+    The ``Z`` suffix is supported on Python versions prior to 3.11, too.
+
+    Args:
+        value: The input data
+        cls: The target type.  Must be :class:`datetime.datetime` or a subclass.
 
     Return:
         The converted datetime instance
 
     Raise:
-        TypeError: If *val* is neither a string nor a datetime
+        TypeError: If *value* is neither a string nor a datetime
+        ValueError: If *value* cannot be parsed as datetime.
     """
     if not isinstance(value, (datetime, str)):
         raise TypeError(
             f"Invalid type {type(value).__name__!r}; expected 'datetime' or " f"'str'."
         )
     if isinstance(value, str):
-        if value[-1] == "Z":
+        if not PY_311 and value[-1] == "Z":
             value = value.replace("Z", "+00:00")
         return cls.fromisoformat(value)
     return value
@@ -555,15 +587,15 @@ _DAYS = "(?P<days>[0-9]+)"
 _HOURS = "(?P<hours>[0-9]+)"
 _MINUTES = "(?P<minutes>[0-9]+)"
 _SECONDS = r"(?P<seconds>[0-9]+)(\.(?P<micros>[0-9]{1,6}))?"
-#: [[[D ]HH:]MM:]SS[.ffffff]
-RE_TIMEDELTA_SIMPLE_1 = re.compile(
-    f"^{_SIGN}((({_DAYS} )?{_HOURS}:)?{_MINUTES}:)?{_SECONDS}$", flags=re.IGNORECASE
+# [[[D ]HH:]MM:]SS[.ffffff]
+RE_TIMEDELTA_SIMPLE = re.compile(
+    f"^{_SIGN}({_DAYS}D,?)?(({_HOURS}:)?{_MINUTES}:)?{_SECONDS}$", flags=re.IGNORECASE
 )
-#: [nnD][nnH][nnM][nnS]
-RE_TIMEDELTA_SIMPLE_2 = re.compile(
+# [nnD][nnH][nnM][nnS]
+RE_TIMEDELTA_SIMPLE_ISO = re.compile(
     f"^{_SIGN}({_DAYS}D)?({_HOURS}H)?({_MINUTES}M)?({_SECONDS}S)?$", flags=re.IGNORECASE
 )
-#: P[nnD][T[nnH][nnM][nnS]]
+# P[nnD][T[nnH][nnM][nnS]]
 RE_TIMEDELTA_ISO = re.compile(
     f"^{_SIGN}"
     r"P(?!\b)"  # "P", but not on a word boundary (e.g., at the end of the string)
@@ -578,14 +610,69 @@ RE_TIMEDELTA_ISO = re.compile(
 )
 
 
-def to_timedelta(value: Union[timedelta, int, float, str], cls: timedelta) -> timedelta:
+def to_timedelta(
+    value: Union[timedelta, int, float, str], cls: Type[timedelta]
+) -> timedelta:
+    """
+    Convert *value* to a :class:`datetime.timedelta`.
+
+    Accepts strings, integers and floats, and timedelta instances.
+
+    Timedelta instances are returned unchanged.
+
+    Integers and floats are interpreted as seconds.
+
+    Supported string formats (all are case-insensitive):
+
+    - :samp:`[±]P[{n}D][T[{nn}H][{nn}M][{nn}[.{ffffff}]S]]` (`ISO durations`_), e.g.:
+
+      - ``P1DT03H04M05S``
+      - ``-P180D``
+      - ``PT4H30M``
+      - ``P1DT30S``
+
+    - :samp:`[±][{nn}D][{nn}H][{nn}M][{nn}S]` (simplified ISO variant), e.g.:
+
+      - ``1d3h4m5s``
+      - ``-180d``
+      - ``4h30m``
+      - ``1d30s``
+
+    - :samp:`[±][{d}D[,]][[{hh}:]{mm}:]{ss}[.{ffffff}]`, e.g.:
+
+      - ``1d,03:04:05``
+      - ``-180D``
+      - ``4:30:00``
+      - ``1d30``
+
+    .. _iso durations: https://en.wikipedia.org/wiki/ISO_8601#Durations
+
+    Args:
+        value: The input data
+        cls: The target type.  Must be :class:`datetime.timedelta` or a subclass.
+
+    Return:
+        The converted timedelta instance
+
+    Raise:
+        TypeError: If *value* is neither a string, float, or int nor a timedelta.
+        ValueError: If *value* cannot be parsed as timedelta.
+
+    .. versionadded:: 24.3.0
+    """
+    if not isinstance(value, (timedelta, str, float, int)):
+        raise TypeError(
+            f"Invalid type {type(value).__name__!r}; expected 'timedelta', 'float', "
+            f"'int', or 'str'."
+        )
+
     if isinstance(value, timedelta):
         return value
 
     if isinstance(value, (int, float)):
         return cls(seconds=value)
 
-    for regex in (RE_TIMEDELTA_ISO, RE_TIMEDELTA_SIMPLE_1, RE_TIMEDELTA_SIMPLE_2):
+    for regex in (RE_TIMEDELTA_ISO, RE_TIMEDELTA_SIMPLE_ISO, RE_TIMEDELTA_SIMPLE):
         match = regex.match(value)
         if match:
             break
@@ -595,14 +682,15 @@ def to_timedelta(value: Union[timedelta, int, float, str], cls: timedelta) -> ti
     parts = match.groupdict(default="0")
     sign_factor = -1 if parts["sign"] == "-" else 1
     days = int(parts["days"]) * sign_factor
-    seconds = int(parts["hours"]) * 3600
-    seconds += int(parts["minutes"]) * 60
-    seconds += int(parts["seconds"])
-    seconds *= sign_factor
+    hours = int(parts["hours"]) * sign_factor
+    minutes = int(parts["minutes"]) * sign_factor
+    seconds = int(parts["seconds"]) * sign_factor
     # Append "0" to "micros" to get a 6-digit number (.7 -> 7 -> 700_000)
     micros = int(f"{parts['micros']:<06}") * sign_factor
 
-    return timedelta(days=days, seconds=seconds, microseconds=micros)
+    return cls(
+        days=days, hours=hours, minutes=minutes, seconds=seconds, microseconds=micros
+    )
 
 
 def to_enum(value: Any, cls: Type[ET]) -> ET: